Run auto-update action more often for validation
This commit updates the `auto-update` action so we run the
script on more types of events.

For every pull request when a python script is updated, we
trigger the action (but skip commiting to master).

Additionally, when any changes to the python script land
on master, the full action is retriggerd.
This way we don't have to wait a week to fix any issues
with the update script.